use non-editable install
